Added window and buffer size flexibility.
Previously, the program was unable to take reshaping/resizing into account because I was rasterizing based on the size of the buffer, instead of the size of the window. These two sizes are now kept in separate variables and are independent. The mouse click function also had to be adapted to take this change into account. In this update I also added full screen functionality and changed some variables names to make their purpose clearer.